requirements.txt
# Core dependencies
requests>=2.31.0
docker>=6.1.0
GitPython>=3.1.40
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# CLI
click>=8.1.7

# Logging
structlog>=23.2.0
